remote_control/requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.0.5
Flask-SocketIO==3.1.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.4.0
python-engineio==3.1.1
python-socketio==2.1.2
six==1.12.0
typing==3.6.6
visitor==0.1.3
Werkzeug==0.14.1
